requirements.txt
 streamlit
 tqdm
 google-generativeai
 jsonschema
 python-dotenv